Refactor: 클래스 분리
예외를 처리하면서 코드가 길어지고, 이름에 대한 기능을 묶은 클래스로 분리하는 것이 맞다고 판단되어 클래스를 분리하게 되엇습니다.
